// front-end/src/app/core/models/user.model.ts
/**
 * Interfaz representa a un usuario tanto en la base de datos
 * como en la aplicacion.
 */
export interface UserInterface {
  uid?: string;
  username: string;
  password: string;
  status: string;
  displayName: string;
  email: string;
  photoUrl: string;

  subjectIds: string[]; // Id de las materia que esta inscrito.
  type: 'student' | 'teacher';
}